--- src/lib/date/format.ts
@@ -1,3 +1,0 @@
-import format from "date-fns/format";
-
-export const formatJpYYYYMD = (date: Date) => format(date, "yyyy年M月d日");
