+++ src/lib/date/index.ts
@@ -1,3 +1,1 @@
-export { formatJpYYYYMD } from "./format";
-export { getSafeDate } from "./getSafeDate";
-export { isValidISODate } from "./is-valid";
+export { formatJpYYYYMD, getSafeDate, isValidISODate } from "./date-fns";
