--- src/lib/date/is-valid.ts
@@ -1,5 +1,0 @@
-import parseISO from "date-fns/parseISO";
-
-export const isValidISODate = (dateString: string) => {
-  return parseISO(dateString).toString() !== "Invalid Date";
-};
